+++ packages/sdk-router/src/sdk.ts
@@ -20,9 +20,8 @@ import {
   SynapseCCTPRouterQuery,
   SynapseRouterQuery,
 } from './utils/types'
+import { TEN_MINUTES, ONE_WEEK, calculateDeadline } from './utils/deadlines'
 import { SynapseCCTPRouter } from './SynapseCCTPRouter'
-const ONE_WEEK_DEADLINE = BigNumber.from(Math.floor(Date.now() / 1000) + 604800) // one week in the future
-const TEN_MIN_DEADLINE = BigNumber.from(Math.floor(Date.now() / 1000) + 600) // ten minutes in the future
 
 type SynapseRouters = {
   [key: number]: SynapseRouter
@@ -311,7 +310,7 @@ class SynapseSDK {
     } else {
       formattedOriginQuery = { ...(originQuery as SynapseRouterQuery) }
     }
-    formattedOriginQuery.deadline = deadline ?? TEN_MIN_DEADLINE
+    formattedOriginQuery.deadline = deadline ?? calculateDeadline(TEN_MINUTES)
 
     let isSwap = false
     if ((destQuery as SynapseCCTPRouterQuery).routerAdapter) {
@@ -320,7 +319,7 @@ class SynapseSDK {
     } else {
       formattedDestQuery = { ...(destQuery as SynapseRouterQuery) }
     }
-    formattedDestQuery.deadline = ONE_WEEK_DEADLINE
+    formattedDestQuery.deadline = calculateDeadline(ONE_WEEK)
 
     let feeAmount!: BigNumber
     let feeConfig!: FeeConfig
@@ -537,6 +536,9 @@ class SynapseSDK {
           quote.maxAmountOut.gt(bestQuote.maxAmountOut ?? BigNumber.from(0))
         ) {
           bestQuote = quote
+          if (bestQuote) {
+            bestQuote.routerAddress = originRouter.routerContract.address
+          }
         }
       } catch (error) {
         console.error(
@@ -667,7 +669,7 @@ class SynapseSDK {
     }
 
     const query = { ...(rawQuery as SynapseRouterQuery) }
-    query.deadline = deadline ?? TEN_MIN_DEADLINE
+    query.deadline = deadline ?? calculateDeadline(TEN_MINUTES)
     const maxAmountOut = query.minAmountOut
 
     return {