+++ packages/synapse-interface/components/StateManagedBridge/InputContainer.tsx
@@ -54,11 +54,16 @@ export const InputContainer = () => {
     0n
 
   const formattedBalance = hasBalances
-    ? formatBigIntToString(fromTokenBalance, fromToken.decimals[fromChainId], 4)
+    ? formatBigIntToString(
+        fromTokenBalance,
+        fromToken?.decimals[fromChainId],
+        4
+      )
     : '0'
 
   useEffect(() => {
     if (
+      fromToken &&
       fromToken.decimals[fromChainId] &&
       stringToBigInt(fromValue, fromToken.decimals[fromChainId]) !== 0n &&
       stringToBigInt(fromValue, fromToken.decimals[fromChainId]) ===
@@ -139,6 +144,7 @@ export const InputContainer = () => {
               focus:border-none
               border-none
               bg-transparent
+              p-0
               max-w-[100px]
               md:max-w-[160px]
               placeholder:text-[#88818C]