+++ packages/synapse-interface/components/StateManagedBridge/OutputContainer.tsx
@@ -75,6 +75,7 @@ export const OutputContainer = ({}) => {
                 focus:ring-0
                 focus:border-none
                 border-none
+                p-0
                 bg-transparent
                 max-w-[190px]
                placeholder:text-[#88818C]