+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.12](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.11...@synapsecns/sdk-router@0.2.12) (2023-08-02)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.2.11](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.10...@synapsecns/sdk-router@0.2.11) (2023-08-02)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.2.10](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.9...@synapsecns/sdk-router@0.2.10) (2023-08-02)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.2.9](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.8...@synapsecns/sdk-router@0.2.9) (2023-08-01)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.2.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.2.7...@synapsecns/sdk-router@0.2.8) (2023-08-01)
 
 **Note:** Version bump only for package @synapsecns/sdk-router